api/requirements.txt
openai
langchain
langchain-openai
langchain-community
fastapi
uvicorn
pyPDF2
#faiss-cpu
tiktoken
pypdf
python-dotenv
pandas
tabulate


